feat(inventario): ingredient cost with decimals
Ingredients measured in g / ml cost a fraction of a peso per unit: a 900 ml
bag of milk at 3.250 COP is 3,61 COP/ml. `ingredientes.cost_cop` (and the two
cost snapshots on movements and count items) were ints, so milk could only be
costed at 4/ml and every recipe with milk came out 11 % too expensive — the
reason recipes carried milk as "0.222 und" instead of "200 ml".

Migration 0036 turns the three columns into numeric(12,2); unitCostFromPack
keeps two decimals; the inventory schema accepts a decimal per-unit cost.
Recipe totals and productos.cost_cop stay whole pesos.

+++ app/(admin)/inventario/actions.ts
@@ -41,7 +41,7 @@ const CreateIngredienteSchema = z
     stockCurrent: z.coerce.number().min(0),
     stockMin: z.coerce.number().min(0),
     mermaPct: z.coerce.number().min(0).max(100).default(0),
-    costCop: z.coerce.number().int().min(0),
+    costCop: z.coerce.number().min(0),
     // Pack-purchase costing (optional). When both are present, cost_cop is
     // derived server-side as round(packCostCop / packQty). Both null/absent
     // means costCop is taken as the manual per-unit cost.
@@ -72,7 +72,7 @@ const UpdateIngredienteSchema = z
     conversionFactor: z.coerce.number().positive().nullable().optional(),
     stockMin: z.coerce.number().min(0),
     mermaPct: z.coerce.number().min(0).max(100).default(0),
-    costCop: z.coerce.number().int().min(0),
+    costCop: z.coerce.number().min(0),
     // Pack-purchase costing (optional). When both are present, cost_cop is
     // derived server-side as round(packCostCop / packQty). Both null/absent
     // means costCop is taken as the manual per-unit cost.
